test version

+++ src/lib.rs
@@ -6,6 +6,7 @@ impl ToCli for String {
     type CliVariant = String;
 }
 
-impl ToCli for u64 {
-    type CliVariant = u64;
+impl ToCli for u128 {
+    type CliVariant = u128;
 }
+
